+++ R/webexercises_default.R
@@ -1,27 +1,34 @@
-#' Create default webex document
+#' Create default webexercises document
 #'
-#' This function wraps \code{rmarkdown::html_document} to configure compilation to embed the default webex CSS and JavaScript files in the resulting HTML.
+#' This function wraps \code{rmarkdown::html_document} to configure
+#' compilation to embed the default webexercises CSS and JavaScript files in
+#' the resulting HTML.
 #'
-#' @details Call this function as the \code{output_format} argument for the \code{\link[rmarkdown]{render}} function when compiling HTML documents from RMarkdown source.
+#' @details Call this function as the \code{output_format} argument
+#'   for the \code{\link[rmarkdown]{render}} function when compiling
+#'   HTML documents from RMarkdown source.
 #' 
-#' @param ... Additional function arguments to pass to \code{\link[rmarkdown]{html_document}}.
+#' @param ... Additional function arguments to pass to
+#'   \code{\link[rmarkdown]{html_document}}.
+#' 
 #' @seealso \code{\link[rmarkdown]{render}}, \code{\link[rmarkdown]{html_document}}
+#' 
 #' @examples
-#' # copy the webex 'R Markdown' template to a temporary file
+#' # copy the webexercises 'R Markdown' template to a temporary file
 #' \dontrun{
 #' my_rmd <- tempfile(fileext = ".Rmd")
-#' rmarkdown::draft(my_rmd, "webex", "webex")
+#' rmarkdown::draft(my_rmd, "webexercises", "webexercises")
 #'
 #' # compile it
-#' rmarkdown::render(my_rmd, webex::webex_default())
+#' rmarkdown::render(my_rmd, webexercises::webexercises_default())
 #' 
 #' # view the result
 #' browseURL(sub("\\.Rmd$", ".html", my_rmd))
 #' }
 #' @export
-webex_default <- function(...) {
-  css <- system.file("reports/default/webex.css", package = "webex")
-  js <- system.file("reports/default/webex.js", package = "webex")
+webexercises_default <- function(...) {
+  css <- system.file("reports/default/webex.css", package = "webexercises")
+  js <- system.file("reports/default/webex.js", package = "webexercises")
 
   knitr::knit_hooks$set(webex.hide = function(before, options, envir) {
     if (before) {