+++ NEWS.md
@@ -1,9 +1,10 @@
-# webex 0.9.3
+# webexercises 1.0.0
 
+* new package name: webexercises
 * updated method to remove smart quotes when rmarkdown >= 2.2
 * new styles for correct and incorrect answers
 * `longmcq()` function for MCQs with long answers (creates a radiobutton interface)
-* `add_webex_to_bookdown()` new function
+* `add_webex_to_bookdown()` new function to add helper functions to bookdown books
 
 # webex 0.9.2